requirements.txt
fastapi
uvicorn
langchain
langchain-groq
groq
openai
python-dotenv
jinja2
aiofiles